Threshold Change and Weekly Tweet Removal
Changed the threshold from a minimum score to a minimum number of
returned tweets. It now defaults to showing 30. Also removed the weekly
list of top tweets. It never showed enough to be useful. Now the script
focuses only on the top 30 tweets from yesterday. Nice and simple and
lets you get on with your life.